all project change

+++ src/main/java/uz/BTService/btservice/common/util/SecurityUtils.java
@@ -29,7 +29,7 @@ public class SecurityUtils {
         return null;
     }
 
-    public static Long getUserId(){
+    public static Integer getUserId(){
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
 
         if(authentication instanceof UsernamePasswordAuthenticationToken token) {
