+++ src/main/java/uz/BTService/btservice/dto/base/BaseServerModifierDto.java
@@ -16,9 +16,9 @@ public class BaseServerModifierDto extends BaseServerDto{
 
     private LocalDateTime updatedDate;
 
-    private Long createdBy;
+    private Integer createdBy;
 
-    private Long modifiedBy;
+    private Integer modifiedBy;
 
     public <DTO extends BaseServerModifierDto, ENTITY extends BaseServerModifierEntity> ENTITY toEntity(DTO dto, ENTITY entity, String... ignoreProperties){
         BeanUtils.copyProperties(dto, entity, ignoreProperties);
