+++ src/main/java/uz/BTService/btservice/repository/CategoryRepository.java
@@ -9,16 +9,16 @@ import uz.BTService.btservice.entity.CategoryEntity;
 import java.util.List;
 import java.util.Optional;
 
-public interface CategoryRepository extends JpaRepository<CategoryEntity, Long> {
+public interface CategoryRepository extends JpaRepository<CategoryEntity, Integer> {
 
     @Query(value = "SELECT btsc.* FROM bts_category btsc WHERE btsc.id=:categoryId AND btsc.status <> 'DELETED'", nativeQuery = true)
-    Optional<CategoryEntity> findByCategoryId(@Param("categoryId") Long categoryId);
+    Optional<CategoryEntity> findByCategoryId(@Param("categoryId") Integer categoryId);
 
     @Query(value = "SELECT btsc.* FROM bts_category btsc WHERE btsc.status <> 'DELETED'", nativeQuery = true)
     List<CategoryEntity> findAllCategory();
 
     @Query(value = "SELECT btsc.* FROM bts_category btsc WHERE btsc.status <> 'DELETED' fetch first :length row only",nativeQuery = true)
-    List<CategoryEntity> findAllById(@Param("length")Long length);
+    List<CategoryEntity> findAllById(@Param("length")Integer length);
 
     @Modifying
     @Query(value = "WITH RECURSIVE sub_category AS (\n" +
@@ -28,7 +28,7 @@ public interface CategoryRepository extends JpaRepository<CategoryEntity, Long> 
             "                            INNER JOIN\n" +
             "                        sub_category parent ON parent.id=child.parent_id\n" +
             ")UPDATE bts_category SET status = 'DELETED' WHERE id IN(SELECT id FROM sub_category)", nativeQuery = true)
-    void categoryDelete(@Param("categoryId") Long categoryId);
+    void categoryDelete(@Param("categoryId") Integer categoryId);
 
     @Query(value = "SELECT btsc.* FROM bts_category btsc WHERE btsc.name=:categoryName",nativeQuery = true)
     Optional<CategoryEntity> findByCreatedByName(@Param("categoryName")String categoryName);
